+++ tests/unit/components/uic-modal-dialog-test.js
@@ -6,7 +6,8 @@ const run = Ember.run;
 moduleForComponent('uic-modal-dialog', 'Unit | Component | uic modal dialog', {
   needs: [
     'component:uic-content-mask',
-    'component:uic-modal'
+    'component:uic-modal',
+    'component:uic-button'
   ],
   unit: true
 });