+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+* Use ButtonComponent instead of button elements in the built in model dialog components.
+
+
+
+
 ### v0.8.0 2016-06-18
 
 * [FEATURE] Use ember-wormhole to enable rendering the `ContextMenuComponent` to a specified DOM element.